+++ src/Media/AudioFilter/DynamicVolBooster.cpp
@@ -55,7 +55,7 @@ void Media::AudioFilter::DynamicVolBooster::GetFormat(AudioFormat *format)
 		Media::AudioFormat fmt;
 		this->sourceAudio->GetFormat(&fmt);
 		format->formatId = 1;
-		format->bitpersample = this->bitCount;
+		format->bitpersample = (UInt16)this->bitCount;
 		format->frequency = fmt.frequency;
 		format->nChannels = fmt.nChannels;
 		format->bitRate = fmt.frequency * fmt.nChannels * this->bitCount;
@@ -71,7 +71,7 @@ void Media::AudioFilter::DynamicVolBooster::GetFormat(AudioFormat *format)
 	}
 }
 
-Int32 Media::AudioFilter::DynamicVolBooster::SeekToTime(Int32 time)
+UInt32 Media::AudioFilter::DynamicVolBooster::SeekToTime(UInt32 time)
 {
 	if (this->sourceAudio)
 	{
